:bento: Add 200m REMA surface DEM mosaic
Scripts to manually download and process the Reference Elevation Model of Antarctica (REMA) v1 geotiff file as part of #8. Had to rename the tiling functions (regular and selective tile). More confusingly is the change of X to Y and Y to X. X is now the low resolution input, and Y is the high resolution groundtruth. W is currently used for REMA, and we may change to using W1, W2, W3 and so on once we add in other relevant datasets #9.