:checkered_flag: Resample MEaSUREs Ice Flow Velocity from 450m to 500m
Because 450m resolution is an odd number to work with. We have 8kmx8km training tiles, and previously the SRCNN would take a 17x17pixel input from MEaSUREs, which is now 16x16pixel instead (i.e. 500m spatial resolution) so that it can be multiplied/divided more easily. Related to #20.