:alembic: Build generator component of SRGAN
Part of #28. Architect the Generator Network component of a Super-Resolution Generative Adversarial Network (SRGAN). The 3 input rasters are now hardcoded convolved to the right shape instead of being resized with nearest neighbour interpolation. The network is much more deeper, consisting of 16 residual blocks, and uses PixelShuffle/depth_to_space instead of Conv2DTranspose for upsampling.

Model was trained for 250 epochs instead of 150 before, and it showed a bit more of a chaotic rise to a PSNR of almost 170 for training set and 150 for validation set. Visual inspection of outputs does not appear to show better perceptual results as before in 607557d60d5b1da68a9d9ab75817e7eb3727e9bc. May need to train for longer and/or attach to discriminator network for it to work better?