:white_check_mark: Using a jupyter notebook to doctest jupyter notebooks
Closes #10. Created a handy jupyter notebook with scripts to parse jupyter notebooks and execute doctests within the function calls! Added a few unit tests as a start to some (but not all) functions in data_prep.ipynb and srgan_train.ipynb. Fits with ease of use by allowing debugging in a jupyter notebook, but not (yet) compatible with code coverage solutions showing you % coverage of your code.

Also note that matplotlib inline magic was removed, as plots still get plotted, and it doesn't work so well (sometimes) when you parse the .ipynb to a .py file with such magics. Created a download_to_path function in data_prep.ipynb as it's hard to create a test for the other functions requiring a big data file as input.