:pushpin: Pin cudatoolkit to come from defaults channel
Build is failing on #37 (docker base image update) because of lack of disk space again. Channel priority in conda is really a mess, so pinning cudatoolkit here to come from defaults/anaconda channel instead of hcc. Effect is a smaller docker image layer ~3GB instead of 5GB. That should keep things within the limits of the Continuous Integration provider.