:hankey: Implement grid single dataset feature in pdal python
First step in breaking down the oneliner pdal bash script into more chunks. Main addition is a processing_pipeline function that uses pdal-python currently, but is in serious need of refactoring to GMT soon. Also added a summary flag to _integration_test_ipynb that is helpful for work in progress behavioural testing.