:checkered_flag: Save raster tile bounds to geojson
Better visibility of the exact areas of the tiles we use for training. Saved in Antarctic Polar Stereographic Projection (EPSG:3031) and  WGS84 (EPSG:4326) for easy rendering in Github. Plots of the tiles uses geopandas instead of shapely now in the jupyter notebook. Also update some README.md files describing some of the new stuff.