:white_check_mark: Add tests for checking 'tile dataset' feature
Ensures that the selective_tiling function works. Not a very well written integration test (there are hard coded elements) but it does capture an important part of the data processing pipeline. The test retrieves the bounding boxes directly from the tiles_3031.geojson file, meaning it excludes the get_window_bounds function (which is itself captured in a unit test only).