:package: Extra 1km padding for misc and lowres tiles
Attempt to increase accuracy of the terrain at the borders of the tile. Widen the raster tile input by 1000m (1km) on each side (left, down, right, up) of the square, i.e. 2000m (2km) padding. That way we don't have to use 'same' padding which pads our tiles with zero (less information and poorer results for the pixels at the edges).