:chart_with_upwards_trend: Reimplement train_dev split on SRGAN training
Becoming a proper data science fellow again. Bring back the train/dev split that was dropped in fdb0464d75dbea6c16b1082e8e84453d472cc1f9 when we moved to an SRGAN architecture in v0.3.0. Using a shuffled 95/5 split for training set and development set respectively, with a seed value of 42 (cultural reference). Metrics plotted using livelossplot once again! Albeit using its 'low-level' API. The metrics are still stored inside a pandas dataframe and converted to a dictionary on the fly at the end of each epoch.

Model has been retrained with 4712/124 tiles in the train/dev set respectively. Metrics on the validation (dev) set does appear very noisy, but it generally follows the trend of the training set metrics so there is no apparent evidence of overfitting (phew!). Next step is to implement crossover analysis on a truly independent test set!