:construction: Chainer implementation of Discriminator Neural Network
As close as possible, a replacement of the original Keras implementation with a Chainer implementation of the discriminator model within the Super Resolution Generative Adversarial Network. Parameter counts down from 6828033 to 6824193, a difference of 3840. This is possibly due to some differences in chainer vs keras's BatchNormalization layer implementation, which I couldn't quite track down (not a simple apples to apples comparison).