:alien: Reimplement PSNR metric calculation in GPU-enabled CuPy
More work on transition to Chainer. Reimplement the Peak Signal-Noise Ratio calculation in CuPy/Numpy instead of Keras. Made several adjustment to the previous implementation that was 'incorrect'. 1) remove the epsilon (fingers crossed that it works later!), 2) output a single value mean squared error (instead of many values per tile), 3) change data range from 2**16 to 2**32 which is for int32. These changes effectively result in an output exactly the same as skimage.measure.compare_psnr, except that it works on a GPU which is helpful as it saves time on copying the arrays back and forth between GPU and CPU memory.