:bulb: Define Adversarial Loss in discriminator and generator loss funcs
Lengthy mathematical definition of a numerically stable Adversarial Loss (Sigmoid Binary Cross Entropy) documented in Markdown/LaTeX. Generator Loss function was fixed to remove division over batchsize (it was already divided, didn't need to do it again), and unit tests added for good measure. Also added the adversarial loss to the generator loss in a 1:1 fashion, and wrote up the code for calculating the standard discriminator loss. Sigmoid activation in Discriminator removed since it is now defined in the loss function.