:memo: Write subsection on Network Conditioning
Mentioning the idea of 'network conditioning', basically adding other sources of information to improve the predictions of a neural network. Tried to explain it in a geospatial context. Explicitly stating that we use concatenation-based conditioning, as it makes the most sense (at least for me) and is used by the pan-sharpening folks. Also cited some DEM super-resolution work.