:memo: Write Methodology subsection on Loss Functions
Defining the various equations forming the Generator and Discriminator of our Generative Adversarial Network (GAN) based on [ESRGAN](https://arxiv.org/abs/1809.00219). Did my best to make the math equations accurate and less intimidating, but there might be some mistakes to iron out. A maybe not so nice fact is that most GAN papers refer to the discriminator's input as 'x', but I've used 'y' as 'x' is already used for the generator's input. Definitely requires some proofreading, but at least this is a start.