:egg: Reproduce some side figures in EGU2019 DeepBedMap poster
Part of #133. Releasing code to reproduce some of the result figures in my DeepBedMap poster at the EGU 2019 General Assembly. The 3D figure uses a test area over Pine Island Glacier that combines the extent of 2007tx, 2010tr and istarxx (instead of just 2007tx before). The elevation error histogram uses only the 2007tx file. Font sizes of the titles, axes labels and so on have been increased, and the resulting figures were saved in a pdf format before being imported to Gravit Designer for poster production.

Also fixed a bug in the bicubic BEDMAP2 figure which had a slightly wrong size before compared to the predicted DeepBedMap image. Graham et al., 2017's synthetic grid was also compared with, but note that I have not actually published the downloading code. The original synthetic NetCDF file can be found at https://doi.org/10.4225/15/57464ADE22F50, and I have converted it to a 16-bit GeoTIFF using xarray.open_dataset, flipping it on the y-axis, and saving it with rasterio.