:construction: Temporarily uninstall PyGMT and skip gmt-required tests
Very hacky workaroud to solve dependabot not wanting to upgrade DeepBedMap's dependencies! Best guess is that the PyGMT VCS dependency is the problem (see #137, #139, #140, #142, #143, #144). So, I'm taking it out of the Pipfile and disabling pygmt required tests, have dependabot upgrade the other dependencies, and then restore pygmt into the Pipfile. Had to add xarray explicitly to Pipfile for the meantime. Onwards!