:alembic: Hyperparameter tuning on ESRGAN with Snow Accumulation input
Here we report on the series of hyperparameter tuning experiments to see what's the best we got with the new W3 Antarctic Snow Accumulation input dataset (hint: it is comparable to our v0.8.0 model, if not slightly better). Kept to using 12 Residual-in-Residual Blocks for our adapted Enhanced Super Resolution Generative Adversarial Network. Primarily tuning the learning rate and number of epochs, with a secondary focus on tuning the residual scaling factor and batch size. Details of the 3 hyperparameter tuning frenzies as below.

1st hyperparameter tuning frenzy achieved best RMSE_test result of 53.32 at https://www.comet.ml/weiji14/deepbedmap/afd927fa856648bebb00551c68c25016 using these hyperparameter settings: learning rate 1e-4, residual scaling 0.15, batch size 64, num_epochs 69. These came from ~100 experimental trials tuned using these range settings:

Learning rate: 1e-3 to 1e-4
Residual scaling: 0.1 to 0.3
Batch size: 64 or 128
Num epochs: 50 to 100

2nd hyperparameter tuning frenzy achieved best RMSE_test result of 43.58 at https://www.comet.ml/weiji14/deepbedmap/0b9b232394da42e394998b112f628696 using these hyperparameter settings: learning rate 7.5e-5, residual scaling 0.15, batch size 128, num_epochs 84. These came ~50 experimental trials tuned using these range settings (Note: continued to use same train.db from 1st run):

Learning rate: 7.5e-4 to 7.5e-5
Residual scaling 0.1 to 0.3
Batch size: 64 or 128
Num epochs 60 to 90

3rd hyperparameter tuning frenzy achieved best RMSE_test result of 45.40 at https://www.comet.ml/weiji14/deepbedmap/908108f973c142d7ba4fba58297b95ea using these hyperparameter settings: learning rate 5e-5, residual scaling 0.15, batch size 128, num_epochs 87. These came from ~50 experimental trials tuned using these range settings (Note: continued to use same train.db from 2nd run):

Learning rate 5e-4 to 5e-5
Residual scaling 0.1 to 0.3
Batch size 64 or 128
Num epochs: 60 to 120

Seems as though a pretty low learning rate works well, even if the number of epochs required is less than 100. I'm a bit skeptical with what the optuna hyperparameter tuning package is doing, as it doesn't seem to sample the hyperparameter search space very well in subsequent tuning frenzies. But overall, the RMSE_test results do stay below 100 in the 2nd and 3rd frenzies so we'll let it be for now.