:truck: Move details on Loss functions and Training to Appendix
The big move! Moving all of the detailed descriptions on our DeepBedMap model's loss functions and training details to Appendix A and Appendix B respectively. Update reference links to equations and some sections, as well as rewording some of the connecting statements at the end of some subsections. Things look a lot cleaner, and I think we're almost ready to put in the big Results!