:sparkles: Code to produce DeepBedMap paper figures
Introducing a new jupyter notebook/python script file to reproduce the figures in our DeepBedMap paper! This initial commit starts with 'Figure 2' (because I can). Basically it is an overview map of the DeepBedMap DEM layered with nice map elements (colorbar and legend) that also includes bounding boxes of our two focus regions Pine Island Glacier and Thwaites Glacier (that will be shown in subsequent figures).

The work was originally in deepbedmap.ipynb (for months now) but things got too messy, and so here we are! Coming up are Figures 1 (DeepBedMap architecture diagram) and 3, 4, 5 (figures showing nice roughness of DeepBedMap).