:boom: Major revamp of DeepBedMap model architecture diagram
Massive work done this week to get the Figure into a state more representative of the actual convolutional neural network model! Thanks to my supervisor Huw and partner Ruzica for their helpful comments on my draft.

I've renamed the tex/pdf files to have fig1_ in front, and the figure now makes use of custom convolutional box shapes from my own PlotNeuralNet fork. Thumbnail images are pulled from the web and have been flattened (instead of 3D perspective as before). The ReLU activation is now represented properly in the boxes (where needed), input/output images and residual-in-residual dense blocks (RRDB) are coloured differently, and I've made a legend/key to clarify what things are! There's also the actual DeepBedMap DEM image on the right side, and bottom labels for the input/core/upsampling module sections. Code has been black linted, and we're now using `subprocess.run` instead of `!` to execute pdflatex.

I've done some work on representing the RRDB blocks but it's too messy and complicated so it's commented out for now. There might be some other details on what's changed in this commit which I've forgotten to mention, but I think this won't be the last tweak to make before submitting the paper!