:speech_balloon: Expand DeepBedMap model arch caption and link clearly
With the refreshed DeepBedMap model architecture diagram at 5b287c506a6f8b92a4bb37489b06e5123dc9656d, we can now reference the figure and parts of the figure a lot easier! Seems to be some duplicated explanations in-line and in- caption but we'll leave it like so for now.