:recycle: Refactor scripts to use PyGMT 0.1.1
Plots should remain pretty much the same, but the code to make them will be vastly simplified since most of the functions (blockmedian and grdview) are now in upstream PyGMT! In data_prep.ipynb, the thin wrapper for `blockmedian` has been removed, and it's good to see that nothing needs to be changed for `grdtrack`. In deepbedmap.ipynb, then thin wrapper for `grdview` has been removed, and I'm tempted to do away with the plot_3d_view function but will keep it for now. Also refactored one line in the thin subplot wrapper to keep DeepCode.ai happy.