:seedling: Include oleron.cpt file
Make it easier to reproduce the DEM figures which use oleron as a colormap. Because my supervisor once said: "It should be a requirement for every paper to be published with the cpt files so that future scientists can reproduce the figures!".

That, and because I didn't using the default 'oleron' cpt in GMT 6.0 because it didn't have a proper hinge at 0. I actually made a note in e15d42428d619faabc941783b2b8dd4c803828e5, that the cpt was taken from https://github.com/GenericMappingTools/gmt/pull/2004/files#diff-e4a3e040bc44aa305d8d1726c511793b. This is basically the oleron cpt in the now released GMT 6.1 at https://github.com/GenericMappingTools/gmt/blob/6.1/share/cpt/oleron.cpt, but without the 'hard hinge' setting so that it works on GMT 6.0.