:ok_hand: LaTeX and Python code to reproduce Author Comments
The LaTeX code for the author comments in reply to the reviewer comments, and the Python code to reproduce the figures, if anyone needs them! This is for 'The Cryosphere', but should be valid for just about any Copernicus Publications journals (I guess). Actual published versions are at https://doi.org/10.5194/tc-2020-74-AC3 and https://doi.org/10.5194/tc-2020-74-AC2 (long story on what happened to AC1). Should have published this earlier but better late than never!