:sparkles: Save discriminator model weights
Saving the discriminator model weights so we can potentially continue training experimental models that are good. The code for this was written a bit hastily, but it works, though I've not found time to change the save_model_weights_and_architecture function to handle the DiscriminatorModel too.

Edited the HyperbandPruner settings to use min_resource=15 and max_resource=150, and after running some experiments, we've stumbled upon a pretty good result with an RMSE_test of 253.26 at https://www.comet.ml/weiji14/deepbedmap/83748fcb506849d78c275d33f8dd3893! Will try to take this model and try to continue training on it, though I really should fix that hole on the left (due to a bad pixel value in the Snow Accumulation grid).