:alembic: Fine tune model on Thwaites test region further
Really pushing the model to work better on the test region! Training was continued from 3ab30305d0c2c6d2e0432679ecdd961fa7765ceb, and the new best result is at https://www.comet.ml/weiji14/deepbedmap/3cd7af0bf68c4ec6a309c6e371501db1 with an RMSE_test of 214.58. Most pixels with negative ice thickness are gone, but there are still some sticking around in the high slope regions. So things are about 99% of the way there, good enough hopefully. For some reason the number of trials don't get to 60 to 90 trials, and so the final table in srgan_train.ipynb is missing some columns and rows, might be something wrong with the HPCs.