:pencil2: Minor revisions to manuscript
If this is a minor revision, I can't imagine what a major one would look like! Mostly dealing with the reviewer's suggested additions/modifications (especially in the Discussion section), and I've refreshed some of the citations too. Split out the author comment citations into a proper bibtex file, because bibitem is a pain. The all-in-one coloured author response and diff was also a bit painful to craft and took a lot of TeX Stackexchange research! Great tools to use are `bibexport`, `latexdiff-git` and `pdfunite`.