:pencil2: Minor revisions in response to Editor's comments
Address a few more comments from the Editor, mostly clarifying a few minor details. Calculated percentage of training area relative to grounded ice regions (in QGIS) to be less than 0.1%, and clarified that DeepBedMap is an alternative rather than an independent method. In terms of styling issues, changed all instances of Figure to Fig. throughout manuscript, and some minor things. Not quite sure what to do about the theta and eta parameters, since they are n-dimensional tensors, neither a scalar or a vector.